+++ src/VaultSharp/V1/SecretsEngines/GoogleCloud/GoogleCloudOAuth2Token.cs
@@ -15,5 +15,23 @@ namespace VaultSharp.V1.SecretsEngines.GoogleCloud
         /// </value>
         [JsonProperty("token")]
         public string Token { get; set; }
+
+        /// <summary>
+        /// Gets or sets expiry time.
+        /// </summary>
+        /// <value>
+        /// The time.
+        /// </value>
+        [JsonProperty("expires_at_seconds")]
+        public long ExpiresAtSeconds { get; set; }
+
+        /// <summary>
+        /// Gets or sets the token time to live.
+        /// </summary>
+        /// <value>
+        /// The ttl.
+        /// </value>
+        [JsonProperty("token_ttl")]
+        public int TokenTimeToLive { get; set; }
     }
 }